refactor: Restructure package for extend
